Add files via upload
In 1.takable_units the program reads two csv files (the available courses and a specific student grades) and based on the these two files specifies all the classes the given student can take the following semester.A student can take a class which she\he has not taken before and has passed (gained at least 50% of the grade) all the prerequisites of it.The classes are sorted based on English alphabet increasingly and are printed in a single line each.
In 2.recommended_classes the program reads the same input files but recommends the next semester classes based on the given student's GPA and list of all grades (read from the csv file). The minimum number of units a student can take are 12 unless there are less units left and if the student's GPA is more than 17/20 she/he can take up to 24 units and if not, she/he can take 20 units at most.